enhanced-fastapi-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.24.1
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
supabase==2.0.0
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
trafilatura==2.0.0
beautifulsoup4==4.12.2
numpy==1.24.3
faiss-cpu==1.7.4
tiktoken==0.5.1
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
openai>=1.6.1
pydantic==2.5.0
python-dotenv==1.0.0
requests-html==0.10.0
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0
lxml>=5.3.0
feedparser==6.0.10
newspaper3k==0.2.8
readability-lxml==0.8.4.1
langchain>=0.0.350
langchain-openai>=0.0.2
chromadb==0.4.18
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
flower==2.0.1 